requirements.txt
flask
reportlab
python-dotenv
requests
